backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
smtplib-async==0.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.2
aiosmtplib==3.0.1
jinja2==3.1.2
google-generativeai==0.3.2
